refactor: 불필요한 추상 클래스 제거
